# docs/integrations/authorization/service-bridge.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Docmost Service Bridge (east-west)",
    "version": "1.1.0",
    "summary": "The east-west endpoints this Docmost fork EXPOSES to an integrating platform.",
...
    "license": { "name": "AGPL-3.0-only", "identifier": "AGPL-3.0-only" },
    "x-canonical": true
  },
  "servers": [
    {
      "url": "https://your-docmost.example",
      "description": "The Docmost fork base URL (e.g. http://localhost:3000 in dev). The global prefix /api is included in the paths below."
    }
  ],
  "security": [{ "serviceSecret": [] }],
...
  "paths": {
...
    "/api/service/spaces": {
      "get": {
        "operationId": "listSpaces",
        "tags": ["spaces"],
        "summary": "List non-personal spaces (with member counts).",
        "description": "Archived spaces are hidden unless includeArchived=true. Scope: spaces:read.",
        "parameters": [{ "name": "includeArchived", "in": "query", "required": false, "schema": { "type": "boolean" } }],
        "responses": {
          "200": { "description": "Spaces in the workspace.", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/SpaceView" } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "503": { "$ref": "#/components/responses/Unconfigured" }
        }
      },
      "post": {
        "operationId": "createSpace",
        "tags": ["spaces"],
        "summary": "Create a space and add the creator as its admin member (atomically).",
        "description": "Both the space row and the creator-admin member row are written in ONE transaction so the outbox projects them together. The creator is resolved from creatorExternalId. Scope: spaces:write.",
        "requestBody": { "required": true, "content": { "application/json": {
          "schema": { "$ref": "#/components/schemas/CreateSpaceRequest" },
          "example": { "name": "Engineering", "creatorExternalId": "ext-abc123" }
        } } },
        "responses": {
          "200": { "description": "The created space.", "content": { "application/json": { "schema": { "type": "object", "required": ["id", "slug", "name"], "properties": { "id": { "type": "string", "format": "uuid" }, "slug": { "type": "string" }, "name": { "type": ["string", "null"] } } } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "503": { "$ref": "#/components/responses/Unconfigured" }
        }
      }
    },
    "/api/service/spaces/{spaceId}": {
      "parameters": [{ "name": "spaceId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
      "get": {
        "operationId": "getSpace",
        "tags": ["spaces"],
        "summary": "Space detail (with member count).",
        "description": "Scope: spaces:read.",
        "responses": {
          "200": { "description": "The space.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SpaceView" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "503": { "$ref": "#/components/responses/Unconfigured" }
        }
      },
      "patch": {
        "operationId": "updateSpace",
        "tags": ["spaces"],
        "summary": "Rename / re-describe a space (slug is stable).",
        "description": "Scope: spaces:write.",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateSpaceRequest" } } } },
        "responses": {
          "200": { "$ref": "#/components/responses/Ok" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "503": { "$ref": "#/components/responses/Unconfigured" }
        }
      }
    },
...
    "/api/service/content/pages/list": {
      "post": {
        "operationId": "listContentPages",
        "tags": ["content"],
        "summary": "Filter-then-retrieve page metadata for an authorized id set.",
        "description": "PRIVILEGED DATA PLANE: returns metadata for EXACTLY the supplied ids (the platform's PDP-authorized set) and does NOT re-authorize. Keyset-ordered by (updated_at desc, id desc); the fork returns up to limit+1 rows so the platform can detect hasMore. Scope: content:read.",
        "requestBody": { "required": true, "content": { "application/json": {
          "schema": { "$ref": "#/components/schemas/ContentListRequest" },
          "example": { "ids": ["33333333-3333-3333-3333-333333333333"], "limit": 50 }
        } } },
        "responses": {
          "200": { "description": "Page summaries (up to limit+1).", "content": { "application/json": { "schema": { "type": "object", "required": ["items"], "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/PublicPageSummary" } } } } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "503": { "$ref": "#/components/responses/Unconfigured" }
        }
      }
    },
    "/api/service/content/spaces/list": {
      "post": {
        "operationId": "listContentSpaces",
        "tags": ["content"],
        "summary": "Filter-then-retrieve space metadata for an authorized id set.",
        "description": "PRIVILEGED DATA PLANE (see listContentPages). Scope: content:read.",
        "requestBody": { "required": true, "content": { "application/json": {
          "schema": { "$ref": "#/components/schemas/ContentListRequest" },
          "example": { "ids": ["55555555-5555-5555-5555-555555555555"], "limit": 50 }
        } } },
        "responses": {
          "200": { "description": "Space summaries (up to limit+1).", "content": { "application/json": { "schema": { "type": "object", "required": ["items"], "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/PublicSpaceSummary" } } } } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "503": { "$ref": "#/components/responses/Unconfigured" }
        }
      }
    },
    "/api/service/content/spaces/{spaceId}": {
      "parameters": [{ "name": "spaceId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
      "get": {
        "operationId": "getContentSpace",
        "tags": ["content"],
        "summary": "A single space's public summary (active only).",
        "description": "Scope: content:read.",
        "responses": {
          "200": { "description": "The space summary.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PublicSpaceSummary" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "503": { "$ref": "#/components/responses/Unconfigured" }
        }
      }
    },
    "/api/collab/force-disconnect": {
      "post": {
        "operationId": "forceDisconnect",
        "tags": ["collab"],
        "summary": "Force-disconnect a user's live collab session on a page after a revocation.",
        "description": "Close the user's live collaboration connections on a page. The fork RE-CHECKS access before disconnecting (fail-safe): if the user can still access the page, it returns { disconnected: false } without disconnecting. Gated by the same x-authz-service-secret header (no per-route scope on this endpoint) plus RemoteOnlyGuard.",
        "requestBody": { "required": true, "content": { "application/json": {
          "schema": { "type": "object", "additionalProperties": false, "required": ["userId", "pageId"], "properties": { "userId": { "type": "string", "format": "uuid" }, "pageId": { "type": "string", "format": "uuid" } } },
          "example": { "userId": "11111111-1111-1111-1111-111111111111", "pageId": "22222222-2222-2222-2222-222222222222" }
        } } },
        "responses": {
          "200": { "description": "Whether the user was disconnected (false if a re-check shows they still have access).", "content": { "application/json": { "schema": { "type": "object", "required": ["disconnected"], "properties": { "disconnected": { "type": "boolean" } } }, "example": { "disconnected": true } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "503": { "$ref": "#/components/responses/Unconfigured" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "serviceSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-authz-service-secret",
        "description": "Shared secret (PLATFORM_AUTHZ_SERVICE_SECRET) sent on every request, compared in constant time."
      }
    },
    "responses": {
      "Ok": {
        "description": "Success (no body of interest).",
        "content": { "application/json": { "schema": { "type": "object", "required": ["ok"], "properties": { "ok": { "type": "boolean" } } } } }
      },
      "BadRequest": {
        "description": "Invalid body (bad shape, unknown key, or a value failing validation).",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Unauthorized": {
        "description": "Missing or wrong x-authz-service-secret.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Forbidden": {
        "description": "The credential lacks the required per-route scope.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "NotFound": {
        "description": "The addressed resource does not exist (or is out of the workspace / archived).",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Conflict": {
        "description": "A uniqueness constraint (e.g. the space slug) is already taken.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "RateLimited": {
        "description": "Per-credential, per-scope rate limit exceeded.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Unconfigured": {
        "description": "No service secret configured on the fork, OR the fork is not in AUTHZ_MODE=remote (fail-closed).",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    },
    "schemas": {
      "ExternalId": {
        "type": "string",
        "pattern": "^[A-Za-z0-9._+-]{1,128}$",
        "description": "The caller's opaque stable identity id; the fork derives the shadow email from it."
      },
...
      "SpaceView": {
        "type": "object",
        "required": ["id", "name", "slug", "description", "visibility", "memberCount", "archived", "createdAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": ["string", "null"] },
          "slug": { "type": "string" },
          "description": { "type": ["string", "null"] },
          "visibility": { "type": "string" },
          "memberCount": { "type": "integer" },
          "archived": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "CreateSpaceRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["name", "creatorExternalId"],
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 120 },
          "slug": { "type": "string", "maxLength": 60 },
          "description": { "type": "string", "maxLength": 1000 },
          "creatorExternalId": { "$ref": "#/components/schemas/ExternalId" }
        }
      },
      "UpdateSpaceRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 120 },
          "description": { "type": "string", "maxLength": 1000 }
        }
      },
...
      "ContentListRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["ids", "limit"],
        "properties": {
          "ids": { "type": "array", "maxItems": 10000, "items": { "type": "string", "format": "uuid" }, "description": "The platform's PDP-authorized id set (the belt). maxItems MUST stay >= the platform's content.maxListAuthorizedIds default (10000) or a well-authorized principal gets a 400; the provider/consumer contract tests enforce this." },
          "spaceId": { "type": "string", "format": "uuid", "description": "Pages only: scope the list to one space." },
          "before": {
            "type": "object",
            "additionalProperties": false,
            "required": ["updatedAt", "id"],
            "properties": { "updatedAt": { "type": "string" }, "id": { "type": "string" } },
            "description": "The decoded keyset cursor position (cursor encode/decode stays on the platform)."
          },
          "limit": { "type": "integer", "minimum": 1, "maximum": 100 }
        }
      },
      "PublicPageSummary": {
        "type": "object",
        "required": ["id", "slugId", "title", "icon", "spaceId", "parentPageId", "position", "createdAt", "updatedAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "slugId": { "type": "string" },
          "title": { "type": ["string", "null"] },
          "icon": { "type": ["string", "null"] },
          "spaceId": { "type": "string", "format": "uuid" },
          "parentPageId": { "type": ["string", "null"] },
          "position": { "type": ["string", "null"] },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "PublicSpaceSummary": {
        "type": "object",
        "required": ["id", "name", "slug", "description", "visibility", "createdAt", "updatedAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": ["string", "null"] },
          "slug": { "type": "string" },
          "description": { "type": ["string", "null"] },
          "visibility": { "type": ["string", "null"] },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Error": {
        "type": "object",
        "description": "Standard Nest exception body; only the HTTP status is contractual.",
        "properties": {
          "statusCode": { "type": "integer" },
          "message": { "type": "string" },
          "error": { "type": "string" }
        }
      }
    }
  }
}